rewrote five.py
